fix: removed console.log,renamed reducer file and trimmed fat of getFilteredByCategory function
